requirements.txt
google-generativeai>=0.3.0
argparse>=1.4.0
streamlit>=1.27.0
beautifulsoup4>=4.9.3
requests>=2.25.1
trafilatura>=1.6.0
lxml>=4.9.0
pandas>=1.5.0
plotly>=5.15.0
